r3216/sofa-dev : FIX: bug on 64 bits a 32bit integer can't be compared with the 64bit std::string::npos as its value is the largest unsigned int, and is out of the 32bits range

+++ modules/sofa/component/visualmodel/OglShader.cpp
@@ -88,8 +88,8 @@ void OglShader::init()
     std::string tempStr = vertFilename.getValue();
     std::string file;
     const std::string SEPARATOR = ";";
-    unsigned int pos = 0;
-    unsigned int oldPos = 0;
+    unsigned long int pos = 0;
+    unsigned long int oldPos = 0;
 
     pos = tempStr.find(SEPARATOR, oldPos);
